python/requirements.txt
Flask==2.1.1
Flask-PyMongo==2.3.0
PyJWT==2.1.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
